Rewritten games module

+++ modules/games/acrophobia.js
@@ -35,7 +35,7 @@ exports.run = (Bastion, message, args) => {
       acronym.push(charPool.random());
     }
 
-    message.channel.sendMessage('', {embed: {
+    message.channel.send({embed: {
       color: Bastion.colors.blue,
       title: 'Acrophobia',
       description: `Game started by ${message.author}. Create a sentence with this acronym: **${acronym.join('. ')}.**`,
@@ -43,15 +43,17 @@ exports.run = (Bastion, message, args) => {
         text: `You have ${2} minutes to make your submission.`
       }
     }}).then(msg => {
-      const collector = msg.channel.createCollector(
+      const collector = msg.channel.createMessageCollector(
         m => !m.author.bot && m.content.split(' ').length == acroLen && matchAcronym(acronym, m.content.split(' ')) && !activeChannels[message.channel.id].usersSubmitted.includes(m.author.id),
         { time: 2 * 60 * 1000 }
       );
-      collector.on('message', (msg, sentences) => {
-        msg.delete().catch(e => {
-          Bastion.log.error(e.stack);
-        });
-        msg.channel.sendMessage('', {embed: {
+      collector.on('collect', (msg, sentences) => {
+        if (msg.deletable) {
+          msg.delete().catch(e => {
+            Bastion.log.error(e.stack);
+          });
+        }
+        msg.channel.send({embed: {
           color: Bastion.colors.dark_grey,
           description: `${msg.author} made their submission.`,
           footer: {
@@ -59,20 +61,24 @@ exports.run = (Bastion, message, args) => {
           }
         }}).then(m => {
           activeChannels[message.channel.id].usersSubmitted.push(msg.author.id);
-          m.delete(5000);
+          m.delete(5000).catch(e => {
+            Bastion.log.error(e.stack);
+          });
         }).catch(e => {
           Bastion.log.error(e.stack);
         });
       });
       collector.on('end', (collection, reason) => {
         if (collection.size == 0) {
-          message.channel.sendMessage('', {embed: {
+          message.channel.send({embed: {
             color: Bastion.colors.red,
             title: 'Acrophobia',
             description: 'Game ended. Unfortunately, no submissions were made for this acronym.'
           }}).then(() => {
             delete activeChannels[message.channel.id];
-            msg.delete();
+            msg.delete().catch(e => {
+              Bastion.log.error(e.stack);
+            });
           }).catch(e => {
             Bastion.log.error(e.stack);
           });
@@ -82,7 +88,7 @@ exports.run = (Bastion, message, args) => {
           for (let i = 0; i < collection.size; i++) {
             submissions.push(`**${i+1}.** ${collection.map(a => a.content)[i]}`);
           }
-          msg.channel.sendMessage('', {embed: {
+          msg.channel.send({embed: {
             color: Bastion.colors.green,
             title: 'Acrophobia',
             description: 'Submissions closed',
@@ -96,16 +102,20 @@ exports.run = (Bastion, message, args) => {
               text: 'Vote by typing the corresponding number of a sentence. You have 60 seconds to vote.'
             }
           }}).then(subMsg => {
-            msg.delete();
-            const votesCollector = msg.channel.createCollector(
+            msg.delete().catch(e => {
+              Bastion.log.error(e.stack);
+            });
+            const votesCollector = msg.channel.createMessageCollector(
               m => !m.author.bot && parseInt(m.content) > 0 && parseInt(m.content) <= collection.size && !activeChannels[message.channel.id].usersVoted.includes(m.author.id),
               { time: 60 * 1000 }
             );
-            votesCollector.on('message', (msg, votes) => {
-              msg.delete().catch(e => {
-                Bastion.log.error(e.stack);
-              });
-              msg.channel.sendMessage('', {embed: {
+            votesCollector.on('collect', (msg, votes) => {
+              if (msg.deletable) {
+                msg.delete().catch(e => {
+                  Bastion.log.error(e.stack);
+                });
+              }
+              msg.channel.send({embed: {
                 color: Bastion.colors.dark_grey,
                 description: `Thank you, ${msg.author}, for voting.`,
                 footer: {
@@ -113,19 +123,23 @@ exports.run = (Bastion, message, args) => {
                 }
               }}).then(m => {
                 activeChannels[message.channel.id].usersVoted.push(msg.author.id);
-                m.delete(5000);
+                m.delete(5000).catch(e => {
+                  Bastion.log.error(e.stack);
+                });
               });
             });
             votesCollector.on('end', votes => {
               if (votes.size == 0) {
-                msg.channel.sendMessage('', {embed: {
+                msg.channel.send({embed: {
                   color: Bastion.colors.red,
                   title: 'Acrophobia',
                   description: 'Game ended. Unfortunately, no votes were given for any submissions.'
                 }}).then(() => {
                   delete activeChannels[message.channel.id].usersSubmitted;
                   delete activeChannels[message.channel.id];
-                  subMsg.delete();
+                  subMsg.delete().catch(e => {
+                    Bastion.log.error(e.stack);
+                  });
                 }).catch(e => {
                   Bastion.log.error(e.stack);
                 });
@@ -137,7 +151,7 @@ exports.run = (Bastion, message, args) => {
                 let count = {};
                 for (let i = 0; i < votes.length; i++) count[votes[i]] = count[votes[i]] ? count[votes[i]]+1 : 1;
                 let winningVoteIndex = Object.keys(count).reduce(function(a, b){ return count[a] > count[b] ? a : b });
-                msg.channel.sendMessage('', {embed: {
+                msg.channel.send({embed: {
                   color: Bastion.colors.blue,
                   title: 'Acrophobia',
                   description: `Game Ended. ${collection.map(s => s.author)[winningVoteIndex - 1]} won by ${count[winningVoteIndex] - 1} of ${votes.length - collection.size} votes`,
@@ -151,7 +165,9 @@ exports.run = (Bastion, message, args) => {
                   delete activeChannels[message.channel.id].usersVoted;
                   delete activeChannels[message.channel.id].usersSubmitted;
                   delete activeChannels[message.channel.id];
-                  subMsg.delete();
+                  subMsg.delete().catch(e => {
+                    Bastion.log.error(e.stack);
+                  });
                 }).catch(e => {
                   Bastion.log.error(e.stack);
                 });
@@ -167,7 +183,7 @@ exports.run = (Bastion, message, args) => {
     });
   }
   else {
-    message.channel.sendMessage('', {embed: {
+    message.channel.send({embed: {
       color: Bastion.colors.red,
       description: `Can\'t start an acrophobia now. Another acrophobia game is already running in this channel.\nPlease wait 3 minutes for it to end.`
     }}).catch(e => {
